Moved the logging init() function out of main Created a namespace logger in logger.h and moved init() and the other global logging functions into it. This patch fixes #2.
